+++ src/test/java/io/fabric8/maven/docker/BuildMojoTest.java
@@ -12,13 +12,16 @@ import io.fabric8.maven.docker.service.BuildService;
 import io.fabric8.maven.docker.service.BuildXService;
 import io.fabric8.maven.docker.service.ImagePullManager;
 import org.apache.maven.plugin.MojoExecutionException;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.MockedConstruction;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
@@ -51,9 +54,20 @@ class BuildMojoTest extends MojoTestBase {
 
     @Mock
     private BuildXService.Exec exec;
+    private MockedConstruction<BuildXService.DockerVersionExternalCommand> dockerVersionExternalCommandMockedConstruction;
 
     private static String getOsDependentBuild(Path buildPath, String docker) {
         return buildPath.resolve(docker).toString().replace('/', File.separatorChar);
+    }
+
+    @BeforeEach
+    void setUp() {
+        dockerVersionExternalCommandMockedConstruction = Mockito.mockConstruction(BuildXService.DockerVersionExternalCommand.class);
+    }
+
+    @AfterEach
+    void tearDown() {
+        dockerVersionExternalCommandMockedConstruction.close();
     }
 
     @Test
@@ -336,6 +350,7 @@ class BuildMojoTest extends MojoTestBase {
 
     private BuildXConfiguration.Builder getBuildXPlatforms(String... platforms) {
         return new BuildXConfiguration.Builder()
+            .nodeName("maven0")
             .platforms(Arrays.asList(platforms));
     }